change project structure
one folder for both client and server
