--- JGitServer/src/main/java/com/jgit/m1/jgitserver/JgitserverApplication.java
@@ -1,13 +1,0 @@
-package com.jgit.m1.jgitserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class JgitserverApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(JgitserverApplication.class, args);
-	}
-
-}
