--- JGitServer/src/test/java/com/jgit/m1/jgitserver/JgitserverApplicationTests.java
@@ -1,13 +1,0 @@
-package com.jgit.m1.jgitserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JgitserverApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
