+++ src/main/java/fr/uca/jgit/JGitServer.java
@@ -1,0 +1,13 @@
+package fr.uca.jgit;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JGitServer {
+
+	public static void main(String[] args) {
+		SpringApplication.run(JGitServer.class, args);
+	}
+
+}
