+++ src/main/java/fr/uca/jgit/command/StateCommit.java
