+++ src/main/java/fr/uca/jgit/model/Folder.java
