+++ src/main/java/fr/uca/jgit/model/JGitObject.java
