+++ src/main/java/fr/uca/jgit/model/Node.java
