+++ src/test/java/fr/uca/jgit/JGitServerTests.java
@@ -1,0 +1,13 @@
+package fr.uca.jgit;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class JGitServerTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
